폴더 구조 변경 #11
app 폴더에는 라우팅과 관련된 페이지, 레이아웃을 넣고
나머지 컴포넌트, 훅, 유틸리티, UI 등은 src/로 이동하도록함.
