--- src/app/components/ui/Button/SubmitButton.tsx
@@ -1,25 +1,0 @@
-import { cn } from "@/app/utils/styleUtils";
-import IconButton from "./IconButton";
-
-interface SubmitButtonProps {
-  label?: string;
-  onClick?: () => void;
-  className?: string;
-}
-
-export default function SubmitButton({
-  label = "추가",
-  onClick,
-  className,
-}: SubmitButtonProps) {
-  return (
-    <IconButton
-      label={label}
-      icon="/icons/plus_black.svg"
-      type="submit"
-      onClick={onClick}
-      className={cn(className, "w-[56px] sm:min-w-[164px]")}
-      showLabelOnMobile={false}
-    />
-  );
-}
