--- src/app/hooks/useGetToDos.ts
@@ -1,13 +1,0 @@
-import { getTodos } from "@/app/api/api";
-import { TODOS_QUERY_KEY } from "@/app/constants";
-import { useSuspenseQuery } from "@tanstack/react-query";
-
-export default function useGetToDos() {
-  const { data, isLoading, isError } = useSuspenseQuery({
-    queryKey: [TODOS_QUERY_KEY],
-    queryFn: () => getTodos(),
-    staleTime: 1000 * 60 * 5,
-  });
-
-  return { data, isLoading, isError };
-}
