+++ src/app/todos/[id]/page.tsx
@@ -1,4 +1,4 @@
-import TodoDetailContent from "@/app/components/todos/TodoDetailContent";
+import TodoDetailContent from "@/components/todos/TodoDetailContent";
 import { Suspense } from "react";
 import { ErrorBoundary } from "react-error-boundary";
