+++ src/constants/queries.ts
@@ -1,0 +1,2 @@
+export const TODOS_QUERY_KEY = "todos";
+export const TODO_DETAIL_QUERY_KEY = "todo_detail";
